// ledger-engine-frontend/src/lib/api/types/api.d.ts
export interface components {
    schemas: {
// ...
        GenerateQrResponse: {
            /** Format: uuid */
            qrCodeId?: string;
            type?: string;
            amount?: string;
            currency?: string;
            description?: string;
            hmacPayload?: string;
            /** Format: date-time */
            createdAt?: string;
            /** Format: date-time */
            expiresAt?: string;
            qrImageBase64?: string;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}